share base RoundTrip
